download function working

+++ networking.h
@@ -2,5 +2,5 @@
  * networking.h
  * networking.c header file */
 
-void serverStart(config_t* Config, short* serverStarted);
+void serverStart(config_t* Config, int* serverStarted);
 void connectionHandler(threadData_t* serverInfo);
